move explosives and refuel to subcomponent
