+++ addons/compat_gm/CfgEventHandlers.hpp
@@ -9,11 +9,3 @@ class Extended_PreInit_EventHandlers {
         init = QUOTE(call COMPILE_SCRIPT(XEH_preInit));
     };
 };
-
-class Extended_InitPost_EventHandlers {
-    class gm_jerrycan {
-        class ADDON {
-            init = QUOTE(call EFUNC(refuel,makeJerryCan));
-        };
-    };
-};
