+++ addons/compat_gm/compat_gm_explosives/CfgAmmo.hpp
@@ -1,25 +1,25 @@
 class CfgAmmo {
     class PipeBombBase;
     class gm_explosive_petn_charge: PipeBombBase {
-        EGVAR(explosive,defuseObjectPosition)[] = {0, 0, 0.02};
+        EGVAR(explosives,defuseObjectPosition)[] = {0, 0, 0.02};
     };
     class gm_explosive_plnp_charge: PipeBombBase {
-        EGVAR(explosive,defuseObjectPosition)[] = {0, 0, 0.02};
+        EGVAR(explosives,defuseObjectPosition)[] = {0, 0, 0.02};
     };
 
     class gm_mine_at_base;
     class gm_mine_at_mn111: gm_mine_at_base {
-        EGVAR(explosive,defuseObjectPosition)[] = {0, 0, 0.02};
+        EGVAR(explosives,defuseObjectPosition)[] = {0, 0, 0.02};
     };
     class gm_mine_at_dm21: gm_mine_at_base {
-        EGVAR(explosive,defuseObjectPosition)[] = {0, 0, 0.02};
+        EGVAR(explosives,defuseObjectPosition)[] = {0, 0, 0.02};
     };
     class gm_mine_at_tm46: gm_mine_at_base {
-        EGVAR(explosive,defuseObjectPosition)[] = {0, 0, 0.02};
+        EGVAR(explosives,defuseObjectPosition)[] = {0, 0, 0.02};
     };
 
     class gm_mine_bounce_base;
     class gm_mine_ap_dm31: gm_mine_bounce_base {
-        EGVAR(explosive,defuseObjectPosition)[] = {0, 0, 0.02};
+        EGVAR(explosives,defuseObjectPosition)[] = {0, 0, 0.02};
     };
 };
