+++ addons/compat_gm/compat_gm_explosives/script_component.hpp
@@ -1,0 +1,5 @@
+#include "\z\ace\addons\compat_gm\script_component.hpp"
+
+#define SUBCOMPONENT explosives
+#undef ADDON
+#define ADDON TRIPLES(PREFIX,COMPONENT,SUBCOMPONENT)
