+++ addons/compat_gm/compat_gm_refuel/CfgEventHandlers.hpp
@@ -1,0 +1,7 @@
+class Extended_InitPost_EventHandlers {
+    class gm_jerrycan {
+        class ADDON {
+            init = QUOTE(call EFUNC(refuel,makeJerryCan));
+        };
+    };
+};
