+++ addons/compat_gm/compat_gm_refuel/CfgVehicles.hpp
@@ -1,0 +1,6 @@
+class CfgVehicles {
+    class gm_ural4320_base;
+    class gm_ural4320_refuel_base: gm_ural4320_base {
+        transportFuel = 0;
+    };
+};
