+++ addons/compat_gm/compat_gm_refuel/config.cpp
@@ -2,21 +2,18 @@
 
 class CfgPatches {
     class ADDON {
-        name = COMPONENT_NAME;
+        addonRootClass = QUOTE(COMPONENT);
         units[] = {};
         weapons[] = {};
         requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = {"ace_csw", "ace_refuel", "gm_core"};
+        requiredAddons[] = {
+            "gm_core",
+            "ace_refuel"
+        };
         skipWhenMissingDependencies = 1;
-        author = ECSTRING(common,ACETeam);
-        authors[] = {"sancron", "nomisum"};
-        url = ECSTRING(main,URL);
         VERSION_CONFIG;
     };
 };
 
-#include "CfgAmmo.hpp"
 #include "CfgEventHandlers.hpp"
-#include "CfgMagazines.hpp"
-#include "CfgWeapons.hpp"
 #include "CfgVehicles.hpp"
